Tests and bugs
Also, implemented Clone for Error, but not sure if I did it correctly.

I would like to implement Consumer with lifetime elision so that we don't have to clone stuff everywhere.
Need to give this idea some more thought. What is the lifetime of each component. Should a topic string exist beyond its call to a client method?