Make PartitionOffset.offset a plain i64
The current behavior of KafkaClient.fetch_group_(topic_)?offsets is to
return early if it encounters any error for any individual partition
fetching, since the only errors returned are for the top-level fetch
request. Consequently, if the top-level Result returned by these methods
is Ok, then so will all the offsets in the underlying PartitionOffset
objects. So wrapping the offset field in a Result just creates
unnecessary encumbrance to the user without providing any benefit.

This PR makes PartitionOffset.offset a plain i64, rather than a
Result<i64>.