Set up integration testing with Docker
This PR sets up a very (very!) basic set up for integration testing this crate. It
takes advantage of Docker, `docker-compose`, and Travis to run integration tests on
the parts of this library that actually need a running Kafka server to execute.
The details are described in #138.

There are still outstanding issues with the approach:

* This set up uses a single-broker Kafka server—to really be useful, the tests
  should run on a multi-broker setup, as this is the more realistic environment that
  production Kafka instances are set up in. I have not yet been able to figure out
  how to get the `wurstmeister/kafka` Docker image to work with a multi-broker setup,
  however.
* Since this crate currently does not have any administrative functionality, like
  creating topics, it relies on creating a fixed topic on the broker before
  executing the tests, which may lead to confusion and potentially the Kafka servers
  getting into weird states while the tests run. To mitigate this, I've made tests
  run serially, rather than in parallel, but this should be improved, if possible.

However, these problems notwithstanding, I think it is still useful to have these
very basic tests and the framework in place so that the most crucial functionality
(querying the Kafka server for metadata, sending messages through a producer, and
receiving messages from a consumer) is tested at least at a basic level, and so
others can begin filling in holes as they work.