Fix off-by-one offset in consumer committing
When the Kafka Consumer commits offsets, the offset that it commits
should be the offset of the next message that it will read.

* https://kafka.apache.org/documentation.html#theconsumer
* https://kafka.apache.org/090/javadoc/org/apache/kafka/clients/consumer/KafkaConsumer.html

However, when the `Consumer` was choosing which offset to send in
the offset commit request inside `Consumer::commit_consumed`, it chose
the offset of the most recently "consumed" message, which is *not* the
next offset that the consumer should read.

This fixes the off-by-one error.