Fix set vs unset in run-all-tests
The bash script was testing for whether it needs to pick default values
by looking at whether the variable was set. However, some variables will
be set to empty explicitly, which causes it to pick the default values
even when we explicitly set a value. Fix this behavior so that it will
not pick defaults when a variable is explicitly set to empty.